requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=1.3.0=pypi_0
annotated-types=0.7.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.31=h06a4308_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.8=pypi_0
diffusers=0.33.0.dev0=pypi_0
docker-pycreds=0.4.0=pypi_0
filelock=3.13.1=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.3.1=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
grpcio=1.70.0=pypi_0
huggingface-hub=0.28.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=8.6.1=pypi_0
jinja2=3.1.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.1.105=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
peft=0.7.0=pypi_0
pillow=11.0.0=pypi_0
pip=25.0=py310h06a4308_0
platformdirs=4.3.6=pypi_0
protobuf=5.29.3=pypi_0
psutil=7.0.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.28.1=pypi_0
safetensors=0.5.2=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.21.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.8.0=py310h06a4308_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
torch=2.2.0+cu121=pypi_0
torchaudio=2.2.0+cu121=pypi_0
torchvision=0.17.0+cu121=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.48.3=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
urllib3=1.26.13=pypi_0
wandb=0.19.6=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xz=5.6.4=h5eee18b_1
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
